app/requirements.txt
aiohttp==3.7.4.post0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
avro-python3==1.10.0
blinker==1.4
certifi==2021.5.30
cffi==1.14.6
Chameleon==3.9.1
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.1
confluent-kafka==1.7.0
cryptography==3.4.8
cssselect==1.1.0
Faker==9.8.3
fastapi==0.68.1
fastapi-chameleon==0.1.12
fastapi-utils==0.2.1
fastavro==1.4.7
greenlet==1.1.1
h11==0.12.0
h2==4.0.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.2
jsonschema==4.2.1
kafka==1.3.5
kafka-python==2.0.2
kaitaistruct==0.9
lxml==4.6.3
mongoengine==0.23.1
motor==2.3.1
multidict==5.1.0
odmantic==0.3.5
protobuf==3.19.1
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
PyJWT==2.3.0
pymongo==3.12.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2021.3
requests==2.26.0
selenium==3.141.0
six==1.16.0
SQLAlchemy==1.4.25
starlette==0.14.2
text-unidecode==1.3
twilio==7.3.1
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
wsproto==1.0.0
yarl==1.6.3
